+++ src/Assertions/UnstructuredHeaderAssertions.php
@@ -2,19 +2,16 @@
 
 namespace KirschbaumDevelopment\MailIntercept\Assertions;
 
+use KirschbaumDevelopment\MailIntercept\AssertableMessage;
 use Symfony\Component\Mime\Email;
 use Symfony\Component\Mime\Header\UnstructuredHeader;
-use KirschbaumDevelopment\MailIntercept\AssertableMessage;
 
 trait UnstructuredHeaderAssertions
 {
     /**
      * Assert unstructured header exists.
-     *
-     * @param string $expected
-     * @param AssertableMessage|Email $mail
      */
-    public function assertMailHasHeader(string $expected, AssertableMessage | Email $mail): void
+    public function assertMailHasHeader(string $expected, AssertableMessage|Email $mail): void
     {
         $this->assertInstanceOf(
             UnstructuredHeader::class,
@@ -25,11 +22,8 @@ trait UnstructuredHeaderAssertions
 
     /**
      * Assert unstructured header exists.
-     *
-     * @param string $expected
-     * @param AssertableMessage|Email $mail
      */
-    public function assertMailMissingHeader(string $expected, AssertableMessage | Email $mail): void
+    public function assertMailMissingHeader(string $expected, AssertableMessage|Email $mail): void
     {
         $this->assertNull(
             $mail->getHeaders()->get($expected),
@@ -39,12 +33,8 @@ trait UnstructuredHeaderAssertions
 
     /**
      * Assert unstructured header is expected value.
-     *
-     * @param string $expected
-     * @param string $expectedValue
-     * @param AssertableMessage|Email $mail
      */
-    public function assertMailHeaderIs(string $expected, string $expectedValue, AssertableMessage | Email $mail): void
+    public function assertMailHeaderIs(string $expected, string $expectedValue, AssertableMessage|Email $mail): void
     {
         $this->assertEquals(
             $expectedValue,
@@ -55,12 +45,8 @@ trait UnstructuredHeaderAssertions
 
     /**
      * Assert unstructured header is not expected value.
-     *
-     * @param string $expected
-     * @param string $expectedValue
-     * @param AssertableMessage|Email $mail
      */
-    public function assertMailHeaderIsNot(string $expected, string $expectedValue, AssertableMessage | Email $mail): void
+    public function assertMailHeaderIsNot(string $expected, string $expectedValue, AssertableMessage|Email $mail): void
     {
         $this->assertNotEquals(
             $expectedValue,